update.py;
_create_experiments() file exists check;

+++ update.py
@@ -86,6 +86,8 @@ class Updater:
         """Create Experiment-{01..12} in self.save_dir"""
         for i in range(1, 13):
             file = os.path.join(self.save_dir, f"Experiment-{i:02}.md")
+            if os.path.exists(file):
+                continue
             with open(file, "w") as file_handle:
                 file_handle.write(f"# Experiment {i}\n\n")
